# docs/swagger.yaml
basePath: /api
definitions:
...
  entity.CardPriority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - PriorityLow
    - PriorityMedium
    - PriorityHigh
  listshttp.cardResponse:
    properties:
      description:
        type: string
      id:
        type: string
      is_archived:
        type: boolean
      list_id:
        type: string
      position:
        type: number
      priority:
        $ref: '#/definitions/entity.CardPriority'
      title:
        type: string
    type: object
  listshttp.listResponse:
    properties:
      board_id:
        type: string
      cards:
        items:
          $ref: '#/definitions/listshttp.cardResponse'
        type: array
      id:
        type: string
      position:
        type: number
      title:
        type: string
    type: object
  services.CardDTO:
    properties:
      description:
        type: string
      id:
        type: string
      is_archived:
        type: boolean
      list_id:
        type: string
      position:
        type: number
      priority:
        $ref: '#/definitions/entity.CardPriority'
      title:
        type: string
    type: object
  services.CreateListDTO:
    properties:
      board_id:
        type: string
      position:
        type: number
      title:
        type: string
    required:
    - board_id
    - title
    type: object
  services.ListDTO:
    properties:
      board_id:
        type: string
      cards:
        items:
          $ref: '#/definitions/services.CardDTO'
        type: array
      created_at:
        type: string
      id:
        type: string
      position:
        type: number
      title:
        type: string
      updated_at:
        type: string
    type: object
  services.UpdateListDTO:
    properties:
      position:
        type: number
      title:
        type: string
    type: object
...
host: localhost:8003
info:
  contact: {}
  description: API for managing kanban boards
  title: Student Kanban API
  version: "1.0"
paths:
...
  /boards/{id}/lists:
    get:
      description: Returns all lists belonging to a given board, each with their cards
      parameters:
      - description: Board UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/listshttp.listResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all lists for a board
      tags:
      - lists
...
  /lists:
    post:
      consumes:
      - application/json
      description: Creates a new list inside a board
      parameters:
      - description: List data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CreateListDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.ListDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new list
      tags:
      - lists
  /lists/{id}:
    delete:
      description: Deletes a list and all its cards by UUID
      parameters:
      - description: List UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a list
      tags:
      - lists
    get:
      description: Returns a single list with its cards
      parameters:
      - description: List UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/listshttp.listResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a list by ID
      tags:
      - lists
    put:
      consumes:
      - application/json
      description: Updates a list's title and/or position
      parameters:
      - description: List UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.UpdateListDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/listshttp.listResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a list
      tags:
      - lists
  /lists/{id}/position:
    patch:
      consumes:
      - application/json
      description: Updates the position of a list within its board
      parameters:
      - description: List UUID
        in: path
        name: id
        required: true
        type: string
      - description: Position data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update list position
      tags:
      - lists
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"